feat: The app interface is 60% completed.
The AppDelegate now contains all the code required to start the App.
Added a dataController to the project.
Extracted all the code related to Layout in a separate extension.
Extracted the code related to the delegate in a separate extension.
Added an extension to UIViewController that contains useful functionalities.
Implemented the required UIElements in the ViewController.
Renamed ViewController to MainViewController.
Reorganized the project structure.